requirements.txt
Markdown==3.6
matplotlib==3.9.0
openai==1.33.0
pandas==2.2.2
PyGithub==2.3.0
python-dotenv==1.0.1
pytz==2024.1
Requests==2.32.3
seaborn==0.13.2
tenacity==8.3.0
tqdm==4.66.4
